// src/global-medical-products/types/medical-product-type.enum.ts
// Type ENUM for Medical Product Type
// This mirrors the Prisma enum MedicalProductType

export enum MedicalProductType {
  antibiotic = 'antibiotic',
  antiparasitic = 'antiparasitic',
  anti_inflammatory = 'anti_inflammatory',
  vitamin = 'vitamin',
  vaccine = 'vaccine',
  anesthetic = 'anesthetic',
  hormone = 'hormone',
  other = 'other',
}